Small refactor to setup merge for ropsten using baked in configs

+++ kiln/devnets/setup.sh
@@ -2,6 +2,7 @@
 # set -e
 
 source parse-args.sh
+source "./fixed.vars"
 source $devnetVars
 
 currentDir=$(pwd)
@@ -22,16 +23,40 @@ then
 fi
 
 
-mkdir $dataDir && mkdir $dataDir/lodestar && mkdir $dataDir/geth && mkdir $dataDir/nethermind && mkdir $dataDir/ethereumjs && mkdir $dataDir/besu && cd $dataDir && git init && git remote add -f origin $setupConfigUrl && git config core.sparseCheckout true && echo "$configGitDir/*" >> .git/info/sparse-checkout && git pull --depth=1 origin main && cd $currentDir
+mkdir $dataDir && mkdir $dataDir/lodestar && mkdir $dataDir/geth && mkdir $dataDir/nethermind && mkdir $dataDir/ethereumjs && mkdir $dataDir/besu
 
-if [ ! -n "$(ls -A $dataDir/$configGitDir)" ] || [ ! -n "$(ls -A $dataDir/$configGitDir/genesis.json)" ] || [ ! -n "$(ls -A $dataDir/$configGitDir/genesis.ssz)" ] || [ ! -n "$(ls -A $dataDir/$configGitDir/nethermind_genesis.json)" ] || [ ! -n "$(ls -A $dataDir/$configGitDir/el_bootnode.txt)" ] || [ ! -n "$(ls -A $dataDir/$configGitDir/bootstrap_nodes.txt)" ]
+if [ -n "$configGitDir" ]
 then
-  echo "Configuration directory not setup properly, remove the data directory and run again."
-  echo "exiting ..."
-  exit;
-else 
-  echo "Configuration discovered!"
+  if [ ! -n "$(ls -A $dataDir/$configGitDir)" ]
+  then
+    cd $dataDir && git init && git remote add -f origin $setupConfigUrl && git config core.sparseCheckout true && echo "$configGitDir/*" >> .git/info/sparse-checkout && git pull --depth=1 origin main && cd $currentDir
+  fi;
+
+  if [ ! -n "$(ls -A $dataDir/$configGitDir)" ] || [ ! -n "$(ls -A $dataDir/$configGitDir/genesis.json)" ] || [ ! -n "$(ls -A $dataDir/$configGitDir/genesis.ssz)" ] || [ ! -n "$(ls -A $dataDir/$configGitDir/nethermind_genesis.json)" ] || [ ! -n "$(ls -A $dataDir/$configGitDir/el_bootnode.txt)" ] || [ ! -n "$(ls -A $dataDir/$configGitDir/bootstrap_nodes.txt)" ]
+  then
+    echo "Configuration directory not setup properly, remove the data directory and run again."
+    echo "exiting ..."
+    exit;
+  else 
+    echo "Configuration discovered!"
+  fi;
+
+  # Load the required variables from the config dir
+  bootNode=$(cat $dataDir/$configGitDir/el_bootnode.txt)
+  bootNode=($bootNode)
+  bootNodeWithSpace=$(IFS=" " ; echo "${bootNode[*]}")
+  bootNode=$(IFS=, ; echo "${bootNode[*]}")
+
+  bootEnr=$(cat $dataDir/$configGitDir/bootstrap_nodes.txt)
+  bootEnr=($bootEnr)
+  bootEnr=$(IFS=" " ; echo "${bootEnr[*]}")
+
+  depositContractDeployBlock=$(cat $dataDir/$configGitDir/deposit_contract_block.txt)
+
+else
+  echo "No configuration specified, assuming the configuration baked in the images and args appropriately set to use it!"
 fi;
+
 
 run_cmd(){
   execCmd=$1;
@@ -72,11 +97,6 @@ then
 fi;
 
 platform=$(uname)
-bootNode=$(cat $dataDir/$configGitDir/el_bootnode.txt)
-bootNode=($bootNode)
-bootNodeWithSpace=$(IFS=" " ; echo "${bootNode[*]}")
-bootNode=$(IFS=, ; echo "${bootNode[*]}")
-
 
 if [ $platform == 'Darwin' ]
 then
@@ -91,12 +111,17 @@ then
   $dockerExec pull $GETH_IMAGE
 
   elName="$DEVNET_NAME-geth"
-  if [ ! -n "$(ls -A $dataDir/geth)" ]
+  if [ ! -n "$(ls -A $dataDir/geth)" ] && [ -n "$configGitDir" ]
   then 
     echo "setting up geth directory"
     $dockerExec run --rm -v $currentDir/$dataDir/$configGitDir:/config -v $currentDir/$dataDir/geth:/data $GETH_IMAGE --datadir /data init /config/genesis.json
   fi;
-  elCmd="$dockerCmd --name $elName $elDockerNetwork -v $currentDir/$dataDir:/data $GETH_IMAGE --datadir /data/geth --authrpc.jwtsecret /data/jwtsecret $GETH_EXTRA_ARGS --bootnodes $EXTRA_BOOTNODES$bootNode"
+
+  elCmd="$dockerCmd --name $elName $elDockerNetwork -v $currentDir/$dataDir:/data $GETH_IMAGE $GETH_EXTRA_ARGS"
+  if [ -n "$configGitDir" ]
+  then
+    elCmd="$elCmd --bootnodes $EXTRA_BOOTNODES$bootNode"
+  fi;
 
 elif [ "$elClient" == "nethermind" ] 
 then
@@ -104,7 +129,14 @@ then
   $dockerExec pull $NETHERMIND_IMAGE
 
   elName="$DEVNET_NAME-nethermind"
-  elCmd="$dockerCmd --name $elName $elDockerNetwork -v $currentDir/$dataDir/$configGitDir:/config -v $currentDir/$dataDir:/data $NETHERMIND_IMAGE --datadir /data/nethermind  --Init.ChainSpecPath=/config/nethermind_genesis.json --JsonRpc.JwtSecretFile /data/jwtsecret $NETHERMIND_EXTRA_ARGS --Discovery.Bootnodes $EXTRA_BOOTNODES$bootNode"
+  elCmd="$dockerCmd --name $elName $elDockerNetwork -v $currentDir/$dataDir:/data"
+  if [ -n "$configGitDir" ]
+  then
+    elCmd="$elCmd -v $currentDir/$dataDir/$configGitDir:/config  $NETHERMIND_IMAGE --Init.ChainSpecPath=/config/nethermind_genesis.json --Discovery.Bootnodes $EXTRA_BOOTNODES$bootNode"
+  else
+    elCmd="$elCmd $NETHERMIND_IMAGE"
+  fi;
+  elCmd="$elCmd $NETHERMIND_EXTRA_ARGS"
 
 elif [ "$elClient" == "ethereumjs" ] 
 then
@@ -112,7 +144,14 @@ then
   $dockerExec pull $ETHEREUMJS_IMAGE
 
   elName="$DEVNET_NAME-ethereumjs"
-  elCmd="$dockerCmd --name $elName $elDockerNetwork -v $currentDir/$dataDir/$configGitDir:/config -v $currentDir/$dataDir:/data $ETHEREUMJS_IMAGE --datadir /data/ethereumjs --gethGenesis /config/genesis.json --jwt-secret /data/jwtsecret $ETHEREUMJS_EXTRA_ARGS --bootnodes=$EXTRA_BOOTNODES$bootNode"
+  elCmd="$dockerCmd --name $elName $elDockerNetwork -v $currentDir/$dataDir:/data"
+  if [ -n "$configGitDir" ]
+  then
+    elCmd="$elCmd -v $currentDir/$dataDir/$configGitDir:/config  $ETHEREUMJS_IMAGE --bootnodes=$EXTRA_BOOTNODES$bootNode"
+  else
+    elCmd="$elCmd $ETHEREUMJS_IMAGE"
+  fi;
+  elCmd="$elCmd --datadir /data/ethereumjs --gethGenesis /config/genesis.json --jwt-secret /data/jwtsecret $ETHEREUMJS_EXTRA_ARGS "
 
 elif [ "$elClient" == "besu" ] 
 then
@@ -120,18 +159,18 @@ then
   $dockerExec pull $BESU_IMAGE
 
   elName="$DEVNET_NAME-besu"
-  elCmd="$dockerCmd --name $elName $elDockerNetwork -v $currentDir/$dataDir/$configGitDir:/config -v $currentDir/$dataDir:/data $BESU_IMAGE --data-path=/data --genesis-file=/config/besu_genesis.json --engine-jwt-secret=/data/jwtsecret $BESU_EXTRA_ARGS --bootnodes=$EXTRA_BOOTNODES$bootNode"
-
+  elCmd="$dockerCmd --name $elName $elDockerNetwork -v $currentDir/$dataDir:/data"
+  if [ -n "$configGitDir" ]
+  then
+    elCmd="$elCmd -v $currentDir/$dataDir/$configGitDir:/config  $BESU_IMAGE --genesis-file=/config/besu_genesis.json --bootnodes=$EXTRA_BOOTNODES$bootNode"
+  else
+    elCmd="$elCmd $BESU_IMAGE"
+  fi;
+  elCmd="$elCmd $BESU_IMAGE --data-path=/data --engine-jwt-secret=/data/jwtsecret $BESU_EXTRA_ARGS"
 fi
 
 echo "lodestarImage: $LODESTAR_IMAGE"
 $dockerExec pull $LODESTAR_IMAGE
-
-bootEnr=$(cat $dataDir/$configGitDir/bootstrap_nodes.txt)
-bootEnr=($bootEnr)
-bootEnr=$(IFS=" " ; echo "${bootEnr[*]}")
-
-depositContractDeployBlock=$(cat $dataDir/$configGitDir/deposit_contract_block.txt)
 
 if [ $platform == 'Darwin' ]
 then
@@ -141,10 +180,26 @@ else
 fi
 
 clName="$DEVNET_NAME-lodestar"
-clCmd="$dockerCmd --name $clName $clDockerNetwork -v $currentDir/$dataDir/$configGitDir:/config -v $currentDir/$dataDir:/data $LODESTAR_IMAGE beacon --rootDir /data/lodestar --paramsFile /config/config.yaml --genesisStateFile /config/genesis.ssz --network.connectToDiscv5Bootnodes --network.discv5.enabled true --eth1.enabled true --eth1.depositContractDeployBlock $depositContractDeployBlock  $LODESTAR_EXTRA_ARGS --bootnodesFile /config/boot_enr.yaml --jwt-secret /data/jwtsecret"
+clCmd="$dockerCmd --name $clName $clDockerNetwork -v $currentDir/$dataDir:/data"
+# mount and use config
+if [ -n "$configGitDir" ]
+then
+  clCmd="$clCmd -v $currentDir/$dataDir/$configGitDir:/config $LODESTAR_IMAGE beacon --paramsFile /config/config.yaml --genesisStateFile /config/genesis.ssz --eth1.depositContractDeployBlock $depositContractDeployBlock --bootnodesFile /config/boot_enr.yaml"
+else
+  clCmd="$clCmd $LODESTAR_IMAGE beacon"
+fi;
+clCmd="$clCmd $LODESTAR_EXTRA_ARGS"
 
 valName="$DEVNET_NAME-validator"
-valCmd="$dockerCmd --name $valName $clDockerNetwork -v $currentDir/$dataDir/$configGitDir:/config -v $currentDir/$dataDir:/data $LODESTAR_IMAGE validator --rootDir /data/lodestar --paramsFile /config/config.yaml $LODESTAR_VALIDATOR_ARGS"
+valCmd="$dockerCmd --name $valName $clDockerNetwork -v $currentDir/$dataDir:/data"
+# mount and use config
+if [ -n "$configGitDir" ]
+then
+  valCmd="$valCmd -v $currentDir/$dataDir/$configGitDir:/config $LODESTAR_IMAGE validator --paramsFile /config/config.yaml"
+else
+  valCmd="$valCmd $LODESTAR_IMAGE validator"
+fi;
+valCmd="$valCmd $LODESTAR_VALIDATOR_ARGS"
 
 echo -n $JWT_SECRET > $dataDir/jwtsecret
 run_cmd "$elCmd"
